created new component for input url playlist and remove codes from mainScreen to SharePlaylistInputBox component
